cleaned up the code and split over several files
